Remove old voting code and extract test to own file
