+++ config/routes.rb
@@ -5,7 +5,6 @@ Rails.application.routes.draw do
     namespace :v1 do
       resources :ideas, only: [:index, :create, :update, :destroy] do
         patch '/vote',   to: "ideas/voter#update"
-        # patch '/downvote', to: "ideas#downvote"
       end
 
       resources :tags, only: [:index]
