+++ test/requests/ideas_endpoints_test.rb
@@ -1,15 +1,6 @@
 require 'test_helper'
 
 class IdeasEndpointTest < ActionDispatch::IntegrationTest
-  def json
-    JSON.parse(response.body)
-  end
-
-  def create_ideas(n)
-    ideas = []
-    n.times { ideas << Idea.create() } && ideas
-  end
-
   test "get ideas by most recent" do
     idea_1, idea_2, idea_3 = create_ideas(3)
     idea_2.update(title: 'bob')
@@ -49,54 +40,6 @@ class IdeasEndpointTest < ActionDispatch::IntegrationTest
     assert_equal 0, Idea.count
   end
 
-  # test "upvote a swill idea" do
-  #   idea = Idea.create()
-  #
-  #   patch "/api/v1/ideas/#{idea.id}/upvote"
-  #
-  #   assert_equal 'plausible', idea.reload.quality
-  # end
-  #
-  # test "upvote a plausible idea" do
-  #   idea = Idea.create(quality: 1)
-  #
-  #   patch "/api/v1/ideas/#{idea.id}/upvote"
-  #
-  #   assert_equal 'genius', idea.reload.quality
-  # end
-  #
-  # test "upvote a genius idea does nothing" do
-  #   idea = Idea.create(quality: 2)
-  #
-  #   patch "/api/v1/ideas/#{idea.id}/upvote"
-  #
-  #   assert_equal 'genius', idea.reload.quality
-  # end
-  #
-  # test "downvote a swill idea does nothing" do
-  #   idea = Idea.create()
-  #
-  #   patch "/api/v1/ideas/#{idea.id}/downvote"
-  #
-  #   assert_equal 'swill', idea.reload.quality
-  # end
-  #
-  # test "downvote a plausible idea" do
-  #   idea = Idea.create(quality: 1)
-  #
-  #   patch "/api/v1/ideas/#{idea.id}/downvote"
-  #
-  #   assert_equal 'swill', idea.reload.quality
-  # end
-  #
-  # test "downvote a genius idea" do
-  #   idea = Idea.create(quality: 2)
-  #
-  #   patch "/api/v1/ideas/#{idea.id}/downvote"
-  #
-  #   assert_equal 'plausible', idea.reload.quality
-  # end
-  #
   test "update an idea" do
     idea = Idea.create()
 
@@ -104,21 +47,5 @@ class IdeasEndpointTest < ActionDispatch::IntegrationTest
 
     assert_equal "title", idea.reload.title
     assert_equal "body",  idea.reload.body
-  end
-
-  test "upvote idea" do
-    idea = Idea.create()
-
-    patch "/api/v1/ideas/#{idea.id}/vote?vote=upvote"
-
-    assert_equal 'plausible', Idea.last.quality
-  end
-
-  test "downvote idea" do
-    idea = Idea.create(quality: 2)
-
-    patch "/api/v1/ideas/#{idea.id}/vote?vote=downvote"
-
-    assert_equal 'plausible', Idea.last.quality
   end
 end
