+++ test/requests/ideas_voter_endpoint_test.rb
@@ -1,0 +1,23 @@
+require 'test_helper'
+
+class IdeasVoterEnpointTest < ActionDispatch::IntegrationTest
+  def json
+    JSON.parse(response.body)
+  end
+
+  test "upvote idea" do
+    idea = Idea.create()
+
+    patch "/api/v1/ideas/#{idea.id}/vote?vote=upvote"
+
+    assert_equal 'plausible', Idea.last.quality
+  end
+
+  test "downvote idea" do
+    idea = Idea.create(quality: 2)
+
+    patch "/api/v1/ideas/#{idea.id}/vote?vote=downvote"
+
+    assert_equal 'plausible', Idea.last.quality
+  end
+end
